Crate aer

A command line tool named “aer” built on the Rust client (“client”) that has an identical scope of a single auraed node.

This tool is for “power-users” and exists as a way of quickly developing and debugging the APIs as we change them. For example an auraed developer might make a change to an API and need a quick way to test the API locally against a single daemon.

Modules

cri

discovery

grpc

observe

runtime

Macros

execute

Executes an rpc call with the default Client and prints the results.

execute_server_streaming

Executes an rpc call with the default Client and prints the results. For use with server streaming requests. The initial response will be printed, followed by printing the stream of messages.